+++ Cliente/src/sdlutils/Game.h
@@ -34,8 +34,8 @@ public:
     void logout();
     void net_thread();
     void game_thread();
-    void crearBala(pair<int,int> pos,double rot);
-    void crearBalaEnemiga(pair<int,int> pos, double rot);
+    void crearBala(pair<int,int> pos,float rot);
+    void crearBalaEnemiga(pair<int,int> pos, float rot);
     
 private:  
     void checkCollision();
